+++ packages/jest-either/test/toMatchLeft.spec.ts
@@ -8,6 +8,10 @@ describe('.toMatchLeft()', () => {
     expect(left({ a: 1, b: 2 })).toMatchLeft({ a: 1 })
   })
 
+  test.skip('passes when object deeply matches the left', () => {
+    expect(left({ a: 1, b: { c: 2, d: 3 } })).toMatchLeft({ b: { c: 2 } })
+  })
+
   test('fails when object does not match the left', () => {
     expect(() =>
       expect(left({ a: 1, b: 2 })).toMatchLeft({ a: 2 })
@@ -28,6 +32,12 @@ describe('.not.toMatchLeft()', () => {
 
   test('passes when object does not match the left', () => {
     expect(left({ a: 1, b: 2 })).not.toMatchLeft({ a: 2 })
+  })
+
+  test.skip('fails when object deeply matches the left', () => {
+    expect(() =>
+      expect(left({ a: 1, b: { c: 2, d: 3 } })).not.toMatchLeft({ b: { c: 2 } })
+    ).toThrowErrorMatchingSnapshot()
   })
 
   test('fails when object matches the left', () => {