+++ packages/jest-either/test/toMatchRight.spec.ts
@@ -8,6 +8,10 @@ describe('.toMatchRight()', () => {
     expect(right({ a: 1, b: 2 })).toMatchRight({ a: 1 })
   })
 
+  test.skip('passes when object deeply matches the right', () => {
+    expect(right({ a: 1, b: { c: 2, d: 3 } })).toMatchRight({ b: { c: 2 } })
+  })
+
   test('fails when object does not match the left', () => {
     expect(() =>
       expect(right({ a: 1 })).toMatchRight({ a: 2 })
@@ -28,6 +32,14 @@ describe('.not.toMatchRight()', () => {
 
   test('passes when object does not match the right', () => {
     expect(right({ a: 1 })).not.toMatchRight({ a: 2 })
+  })
+
+  test.skip('fails when object deeply matches the right', () => {
+    expect(() =>
+      expect(right({ a: 1, b: { c: 2, d: 3 } })).not.toMatchRight({
+        b: { c: 2 }
+      })
+    ).toThrowErrorMatchingSnapshot()
   })
 
   test('fails when object matches the right', () => {